+++ packages/core/src/canvas/fills.ts
@@ -7,7 +7,8 @@ export function drawNodeFill(
   canvas: Canvas,
   node: SceneNode,
   rect: Float32Array,
-  hasRadius: boolean
+  hasRadius: boolean,
+  fill?: Fill
 ): void {
   switch (node.type) {
     case 'VECTOR': {
@@ -30,7 +31,7 @@ export function drawNodeFill(
       }
       break
     case 'TEXT':
-      r.renderText(canvas, node)
+      r.renderText(canvas, node, fill)
       break
     case 'LINE':
       canvas.drawLine(0, 0, node.width, node.height, r.fillPaint)